requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
pyVinted==0.5.2
requests==2.32.3
urllib3==2.3.0
yarl==1.18.3
